# sigPtrGetZIn32 returns a struct tuple