In a trace from here: https://developercommunity.visualstudio.com/content/problem/1035124/trace-of-editing-experience-with-in-memory-cross-p-1.html

I found what looked like needless allocations of a tuple of two integers:

![image](https://user-images.githubusercontent.com/6309070/82163409-c903b580-985f-11ea-9a62-5d379c978d39.png)

Unfortunately this one is a few lines of code since it's called a lot within this file. It does catch this one though, too:

![image](https://user-images.githubusercontent.com/6309070/82163906-a8892a80-9862-11ea-9189-abed27138800.png)