@TIHan 2020-05-26
@cartermp try to use this benchmark test, https://github.com/dotnet/fsharp/blob/master/benchmarks/CompilerServiceBenchmarks/Program.fs#L194, to see if it improves it.