@TIHan 2020-05-27
@KevinRansom , I just need to know if it actually does improve CPU time. I think it will, but because the struct is bigger than 8 bytes, I only want to be sure. There is a 16 byte rule of thumb I go by, but it's just that, a rule of thumb.