@KevinRansom 2020-05-28
@TIHan ,  the rationale for this is similar to the value option argument.  struct tuple was not available when the author wrote this code. It is likely the developer of this functionality would have used struct option, for sure we see it in PRs for core functionality more frequently these days.  As before my intuition is that this may have a benefit but that benefit might be very small and difficult to measure reliably.   My preference is to take this change since it doesn't appear to change things semantically, and may reduce allocations slightly.