@cartermp 2020-05-29
``` ini

BenchmarkDotNet=v0.11.3, OS=Windows 10.0.19041
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit LegacyJIT-v4.8.4180.0 DEBUG
  Job-FEFVFA : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4180.0

InvocationCount=1  UnrollFactor=1  
```

Before:
|                                 Method |        Mean |     Error |    StdDev |       Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------------------------- |------------:|----------:|----------:|-------------:|------------:|------------:|------------:|--------------------:|
|                   ParsingTypeCheckerFs |   208.22 ms |  4.152 ms |  9.200 ms |   208.376 ms |   7000.0000 |   2000.0000 |           - |            42.76 MB |
|                              ILReading | 1,545.98 ms | 29.268 ms | 30.056 ms | 1,556.268 ms | 193000.0000 |  63000.0000 |   5000.0000 |          1117.37 MB |
| TypeCheckFileWith100ReferencedProjects |    12.31 ms |  1.552 ms |  4.576 ms |     9.669 ms |   1000.0000 |   1000.0000 |           - |             7.38 MB |

After:

|                                 Method |        Mean |     Error |    StdDev |       Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------------------------- |------------:|----------:|----------:|-------------:|------------:|------------:|------------:|--------------------:|
|                   ParsingTypeCheckerFs |   191.39 ms |  1.893 ms |  1.678 ms |   190.873 ms |   7000.0000 |   2000.0000 |           - |            42.77 MB |
|                              ILReading | 1,529.46 ms | 30.370 ms | 49.041 ms | 1,527.549 ms | 191000.0000 |  63000.0000 |   6000.0000 |          1112.69 MB |
| TypeCheckFileWith100ReferencedProjects |    13.11 ms |  1.543 ms |  4.550 ms |     9.731 ms |   1000.0000 |   1000.0000 |           - |             7.38 MB |

As expected, this marginally reduces some allocations and CPU time in IL reading